fix: Add direct ref for data.stack-utils
As of now, we are introducing a direct reference to the main branch of the
data.stack-utils repository. This change is necessary to incorporate and test
new features in data.stack-utils. After thorough testing, we will release the
latest changes and update the package.json with the appropriate release
reference.